+++ tests/test_magic.py
@@ -33,9 +33,9 @@ class LineTodo(MagicTodoMixin):
     ident: str = "LINE TODO"
 
 
-@params("line", ["foo bar baz", "test | foo bar baz"])
+@params("line", ["1900-01-01 foo bar baz", "test | 1900-01-01 foo bar baz"])
 def test_line_todo(line: str) -> None:
     """Test that to_line and from_line spells work."""
     todo = LineTodo.from_line(line).unwrap()
     assert todo.desc == "foo bar baz"
-    assert todo.to_line() == "test | foo bar baz"
+    assert todo.to_line() == "test | 1900-01-01 foo bar baz"